requirements.txt
numpy~=1.21.2
pandas~=1.3.3
pillow~=8.3.2
pytest~=6.2.5